Remove openapi folder from routes/balances
- Entities at the route level should be under `/entities`
- A separate OpenApi entity should not be required – we assume that route level entities have OpenApi related decorators
