--- src/routes/balances/openapi/api-balance.ts
@@ -1,14 +1,0 @@
-import { Balance as DomainBalance } from '../entities/balance.entity';
-import { ApiProperty } from '@nestjs/swagger';
-import { TokenInfo as ApiTokenInfo } from './api-token-info';
-
-export class Balance implements DomainBalance {
-  @ApiProperty()
-  balance: string;
-  @ApiProperty()
-  fiatBalance: number;
-  @ApiProperty()
-  fiatConversion: number;
-  @ApiProperty()
-  tokenInfo: ApiTokenInfo;
-}
