--- src/routes/balances/openapi/api-token-info.ts
@@ -1,18 +1,0 @@
-import { TokenInfo as DomainTokenInfo } from '../entities/token-info.entity';
-import { ApiProperty } from '@nestjs/swagger';
-import { TokenType } from '../entities/token-type.entity';
-
-export class TokenInfo implements DomainTokenInfo {
-  @ApiProperty()
-  address: string;
-  @ApiProperty()
-  decimals: number;
-  @ApiProperty()
-  logoUri: string;
-  @ApiProperty()
-  name: string;
-  @ApiProperty()
-  symbol: string;
-  @ApiProperty({ enum: Object.values(TokenType) })
-  tokenType: TokenType;
-}
